rename package names of spring boot starter
