+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-boot-starter/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -19,7 +19,7 @@
   "groups": [
     {
       "name": "spring.shardingsphere.datasource",
-      "type": "org.apache.shardingsphere.jdbc.spring.boot.SpringBootConfiguration"
+      "type": "org.apache.shardingsphere.driver.spring.boot.SpringBootConfiguration"
     },
     {
       "sourceType": "org.apache.shardingsphere.sharding.yaml.config.YamlShardingStrategyConfiguration",
@@ -72,13 +72,13 @@
     {
       "name": "spring.shardingsphere.props",
       "type": "java.util.Properties",
-      "sourceType": "org.apache.shardingsphere.jdbc.spring.boot.prop.SpringBootPropertiesConfigurationProperties",
+      "sourceType": "org.apache.shardingsphere.driver.spring.boot.prop.SpringBootPropertiesConfigurationProperties",
       "description": "Customize shardingsphere properties."
     },
     {
       "name": "spring.shardingsphere.datasource.names",
       "type": "java.lang.String",
-      "sourceType": "org.apache.shardingsphere.jdbc.spring.boot.SpringBootConfiguration"
+      "sourceType": "org.apache.shardingsphere.driver.spring.boot.SpringBootConfiguration"
     },
     {
       "name": "spring.shardingsphere.datasource",
