+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-boot-starter/src/test/resources/META-INF/services/org.apache.shardingsphere.encrypt.strategy.spi.Encryptor
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.jdbc.spring.boot.fixture.TestEncryptor
+org.apache.shardingsphere.driver.spring.boot.fixture.TestEncryptor
