+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-orchestration-spring/shardingsphere-jdbc-orchestration-spring-boot-starter/src/main/resources/META-INF/spring.provides
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-provides: sharding-jdbc-orchestration-spring-boot-starter
+provides: shardingsphere-jdbc-orchestration-spring-boot-starter
